# generated-sources/swagger.json
{"openapi":"3.0.1","info":{"title":"Pruduct Management System API","description":"This API exposes endpoints to manage employees.","contact":{"name":"Mustafa Akocak","email":"your.email@gmail.com"},"version":"1.0"},"servers":[{"url":"http://localhost:5173","description":"Development"}],"paths":{"/api/v1/product":{"put":{"tags":["product-controller"],"operationId":"insert","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}},"post":{"tags":["product-controller"],"operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}}},"/api/v1/product/type":{"put":{"tags":["product-controller"],"operationId":"addProductType","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductTypeDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductTypeDTO"}}}}}}},"/api/v1/product/type/list":{"post":{"tags":["product-controller"],"operationId":"listAllProductType","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductTypeDTO"}}}}}}}},"/api/v1/product/list":{"post":{"tags":["product-controller"],"operationId":"listAll","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}}}}}},"/api/v1/product/{id}":{"get":{"tags":["product-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}},"delete":{"tags":["product-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}}},"/api/v1/product/type/{id}":{"get":{"tags":["product-controller"],"operationId":"getProductType","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductTypeDTO"}}}}}}}},"components":{"schemas":{"CategoryDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"productDTO":{"$ref":"#/components/schemas/ProductDTO"}}},"ProductDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"categoryDTOList":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDTO"}}}},"ProductTypeDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"}}}}}}